unary now separate and has type_check function
