+++ src/ast/expr/mod.rs
@@ -1,7 +1,9 @@
 mod expr;
 mod literal;
 mod operator;
+mod unary;
 
 pub use expr::*;
 pub use literal::*;
 pub use operator::*;
+pub use unary::*;
